services/docker-requirements.txt
celery[redis]
psycopg2-binary
boto3
python-dotenv
pdfplumber
python-docx
transformers
torch